+++ src/main/java/racingcar/comparable/NumberComparable.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.comparable;
 
 public interface NumberComparable<T extends Number> {
     int compare(T x, T y);
