+++ src/main/java/racingcar/random/RandomIntegerGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.random;
 
 import camp.nextstep.edu.missionutils.Randoms;
