+++ src/main/java/racingcar/random/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.random;
 
 public interface RandomNumberGenerator<T extends Number> {
     T pickNumber();
